web/requirements.txt
aiofiles
fastapi[standard]
uvicorn
pytest
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
hivemind @ git+https://github.com/learning-at-home/hivemind@10f82ee9e080c4fdbc95f2f26971a7afd55f43fa
web3